google_places_searcher/requirements.txt
psycopg2-binary==2.8.4
pymongo==3.9.0
requests==2.23.0

